refactor: combine partner and challenges section
