--- apps/evm/src/pages/Fusion/components/Challenges/index.tsx
@@ -1,1 +1,0 @@
-export { Challenges } from './Challenges';
