+++ apps/evm/src/pages/Fusion/components/PartnersAndChallenges/PartnerCard.style.tsx
