+++ apps/evm/src/pages/Fusion/components/PartnersAndChallenges/PartnerCard.tsx
