+++ apps/evm/src/pages/Fusion/components/PartnersAndChallenges/index.tsx
@@ -1,0 +1,1 @@
+export { PartnersAndChallenges } from './PartnersAndChallenges';
