--- apps/evm/src/pages/Fusion/components/PartnersSection/index.tsx
@@ -1,1 +1,0 @@
-export { PartnersSection } from './PartnerSection';
